HMW-107 Fixed an issue of layers showing up in the legend, even though the map zoom is out of range.
